requirements.txt
django==5.1.5
Pillow==11.1.0
drf_yasg==1.21.8
python-dotenv==1.0.1
djangorestframework==3.15.2
